refactor(mocks): extract memory module from mockApi (#991 M3)
Move memories state, OwnedMemory/stripMemoryOwner, and listMemory/deleteMemory into memory.ts as memoryApi, spread into index.ts.
